fixbug linter
Former-commit-id: 0023ca64f11822b3d049151b6900d11915e21085
Former-commit-id: 33a979910dc88cfdf89d386263e4f40e6e15ed9b

+++ types/tx.go
@@ -77,7 +77,7 @@ func (txgroup *Transactions) Check(minfee int64) error {
 	for i := 1; i < len(txs); i++ {
 		totalfee += txs[i].GetFee()
 	}
-	txSize := int64(Size(txs[0]))
+	txSize := Size(txs[0])
 	realFee := int64(txSize/1000+1) * minfee
 	if txs[0].Fee < realFee+totalfee {
 		return ErrTxFeeTooLow
